requirements.txt
aiofiles==0.7.0
anyio==3.3.1
argcomplete==1.10.3
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
beautifulsoup4==4.8.2
bibtexparser==1.2.0
bleach==3.2.1
blis==0.7.4
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
ci-info==0.2.0
click==8.0.1
compressed-rtf==1.0.6
configobj==5.0.6
configparser==5.0.2
cycler==0.10.0
cymem==2.0.5
de-core-news-sm==2.3.0
decorator==4.4.2
defusedxml==0.6.0
docx2txt==0.8
ebcdic==1.1.1
EbookLib==0.17.1
entrypoints==0.3
etelemetry==0.2.2
extract-msg==0.28.7
filelock==3.0.12
frontend==0.0.3
future==0.18.2
fuzzywuzzy==0.18.0
gensim==3.8.3
guess-language-spirit==0.5.3
h11==0.12.0
httplib2==0.19.1
idna==2.10
IMAPClient==2.1.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.2.0
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
murmurhash==1.0.5
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.1
networkx==2.6.3
nibabel==3.2.1
nipype==1.6.1
nltk==3.6.2
notebook==6.1.4
numpy==1.19.5
olefile==0.46
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
pathlib==1.0.1
pdfminer.six==20191110
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.0
pl-core-news-sm==2.3.0
plac==1.1.3
preshed==3.0.5
prometheus-client==0.8.0
prompt-toolkit==3.0.8
prov==2.0.0
ptyprocess==0.6.0
pycparser==2.20
pycryptodome==3.10.1
pydot==1.4.2
Pygments==2.7.1
PyMuPDF==1.18.18
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-Levenshtein==0.12.2
python-pptx==0.6.19
pytz==2020.1
pyxnat==1.4
pyzmq==19.0.2
qgrid==1.3.1
qtconsole==4.7.7
QtPy==1.9.0
rdflib==6.0.0
regex==2021.8.28
requests==2.24.0
scikit-learn==0.23.2
scipy==1.6.0
seaborn==0.11.0
Send2Trash==1.5.0
simplejson==3.17.5
six==1.16.0
sklearn==0.0
smart-open==3.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
spacy==2.3.5
SpeechRecognition==3.8.1
srsly==1.0.5
starlette==0.16.0
stop-words==2018.7.23
terminado==0.9.1
testpath==0.4.4
textract==1.6.4
thinc==7.4.5
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.56.0
traitlets==5.0.5
traits==6.2.0
tzlocal==3.0
urllib3==1.25.10
uvicorn==0.15.0
wasabi==0.8.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==1.2.0
XlsxWriter==3.0.1
